requirements.txt
discord.py
discord-webhook
requests